[fei4220.1.changesets] Setup publishing and force Node 16 or above
