--- lerna.json
@@ -1,18 +1,0 @@
-{
-    "packages": [
-        "packages/*"
-    ],
-    "command": {
-        "publish": {
-            "ignoreChanges": [
-                "*.test.js",
-                "*.vite_test.js",
-                "*.md",
-                "*.js.snap"
-            ]
-        }
-    },
-    "version": "independent",
-    "npmClient": "yarn",
-    "useWorkspaces": true
-}
