+++ package.json
@@ -10,12 +10,16 @@
     "packages/*",
     "build-settings"
   ],
+  "engines": {
+    "node": ">=16"
+  },
   "devDependencies": {
     "@babel/core": "^7.16.12",
     "@babel/eslint-parser": "^7.16.5",
     "@babel/eslint-plugin": "^7.16.5",
     "@babel/preset-env": "^7.16.11",
     "@babel/preset-flow": "^7.16.7",
+    "@changesets/cli": "^2.21.0",
     "@google-cloud/debug-agent": "^5.2.8",
     "@google-cloud/logging-winston": "^4.1.1",
     "@google-cloud/profiler": "^4.1.7",
@@ -47,7 +51,6 @@
     "jest": "^27.2.4",
     "jest-extended": "^2.0.0",
     "jest-when": "^3.4.1",
-    "lerna": "^4.0.0",
     "prettier": "^2.5.0",
     "rollup": "^2.57.0",
     "rollup-plugin-auto-external": "^2.0.0",
@@ -63,14 +66,9 @@
     "clean": "rm -rf packages/wonder-stuff-*/dist && rm -rf packages/wonder-stuff-*/node_modules",
     "coverage": "yarn run jest --coverage",
     "flow:ci": "flow check --max-workers 0",
-    "dry-run-publish": "yarn run publishchecks && lerna version --no-git-tag-version --no-push",
     "lint": "eslint --config .eslintrc.js packages",
     "lint:ci": "eslint --config .eslintrc.js",
-    "publishchecks": "git diff --stat --exit-code HEAD && yarn build && node utils/pre-publish-check.js",
-    "publish": "yarn run publishchecks && lerna publish",
+    "publish:ci": "node utils/publish/pre-publish-check-ci.js && git diff --stat --exit-code HEAD && yarn build && changeset publish",
     "test": "yarn jest"
-  },
-  "dependencies": {
-    "@changesets/cli": "^2.21.0"
   }
 }
